Handle mouse grabs better. Add outline function support. Lazy redraws.
Lots more code to deal with mouse grabbing and multiple buttons.

Switched from pure rectangle extents for widgets to paths,
but using a rectangle to optimize hit tests.

Made redraw happen only when the event queue is empty. This required a
separate event reading thread so that pending events could be noticed.