Manage pixmap size in nickle thread instead of gtk thread.
Executing the gtk resize callback could overlap with rendering
causing partial results to appear in the output where rendering was sent to
a pixmap in the process of being destroyed. This eliminates all use of gdk
pixmaps and instead uses X pixmaps to avoid using gdk APIs from the nickle
thread.